+++ src/App.tsx
@@ -1,34 +1,25 @@
-import { css } from '@emotion/react';
-
-import { Button } from './common/components/button';
+import { KakaoAuthButton } from './common/components/kakaoauthbutton';
+import { BaseButton } from './common/components/systembutton';
 
 function App() {
   return (
     <div>
-      <Button>Default</Button>
+      <BaseButton>Default</BaseButton>
 
-      <Button variant="destructive" size="sm">
+      {/* 디자인 시스템 테마 주입 */}
+      <BaseButton variant="destructive" size="sm">
         Small + Destructive
-      </Button>
+      </BaseButton>
 
-      <Button
-        variant="outline"
-        size="lg"
-        cssProps={css`
-          background-color: gray;
-          border: 10px solid green;
-          font-size: 1.25rem;
-        `}
-        onClick={() => console.log('clicked!')}
-      >
-        Large +Outline + Override
-      </Button>
-
-      <Button variant="link" asChild>
+      {/* asChild에 따른 Slot 사용 */}
+      <BaseButton variant="link" asChild>
         <a href="https://example.com" target="_blank" rel="noreferrer">
           External Link
         </a>
-      </Button>
+      </BaseButton>
+
+      {/* 내부적으로 고정된 버튼 스타일 및 이벤트 사용 */}
+      <KakaoAuthButton>카카오 로그인</KakaoAuthButton>
     </div>
   );
 }
