+++ src/common/components/kakaoauthbutton/kakaoAuthButton.styles.ts
@@ -1,0 +1,10 @@
+// styles/base-button.styles.ts
+import { css } from '@emotion/react';
+
+export const kakaoButtonStyle = css`
+  width: 100px;
+  height: 20px;
+  color: black;
+  background-color: yellow;
+  border-radius: 10px;
+`;
